requirements.txt
Flask==2.3.2
Flask_Login==0.6.2
Flask_Migrate==4.0.4
flask_sqlalchemy==3.0.5
Flask_WTF==1.1.1
python-dotenv==1.0.0
SQLAlchemy==1.4.39
Werkzeug==2.3.6
WTForms==3.0.1
psycopg2-binary==2.9.6
email-validator==2.0.0.post2
